+++ src/utils/defineInput.ts
@@ -1,4 +1,4 @@
-import { FunctionalComponent, h } from 'vue';
+import { FunctionalComponent, VNode, h } from 'vue';
 
 export interface InputProps extends Object {
   modelValue: string | boolean;
@@ -23,13 +23,26 @@ export function defineInput<Props>(
     { slots, emit }
   ) => {
     const { modelValue, ...restOfProps } = props;
+
+    // @TODO hack to support CamelCase Ionic events
+    const onVnodeBeforeMount = (vnode: VNode) => {
+      if (vnode.el) {
+        vnode.el.addEventListener(updateEvent.replace('onIon', 'ion'), (e: Event) => {
+          emit(InputEvents.onUpdate, (e?.target as any)[modelProp]);
+        });
+        for (const [key, prop] of Object.entries(props)) {
+          if (key.startsWith('onIon')) {
+            vnode.el.addEventListener(key.replace('onIon', 'ion'), prop);
+          }
+        }
+      }
+    };
     return h(
       ionTag,
       {
         ...restOfProps,
-        [modelProp]: modelValue,
-        [updateEvent.replace('Ion', '')]: (e: Event) =>
-          emit(InputEvents.onUpdate, (e?.target as any)[modelProp])
+        onVnodeBeforeMount,
+        [modelProp]: props.hasOwnProperty('modelValue') ? modelValue : (props as any)[modelProp],
       },
       slots.default && slots.default()
     );